% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_statistics.R
\name{addClusterMembership}
\alias{addClusterMembership}
\title{Add information about connected subcomponents of the peptide-protein graph
to quantitative data.}
\usage{
addClusterMembership(
  quantification_data,
  peptide_protein_graph,
  protein_column = "ProteinName"
)
}
\arguments{
\item{quantification_data}{MS data, preferably in `MSstats` or `MSstatsTMT` format.}

\item{peptide_protein_graph}{graph created by the `createPeptideProteinGraph` function.}

\item{protein_column}{name of a column with protein names.}
}
\description{
Add information about connected subcomponents of the peptide-protein graph
to quantitative data.
}
