% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_assignments.R
\name{adjustProteinAssignments}
\alias{adjustProteinAssignments}
\title{Add missing proteins to PSM-level data based on a fasta database}
\usage{
adjustProteinAssignments(
  quantification_data,
  fasta_db_path,
  protein_column = "ProteinName",
  peptide_column = "PeptideSequence",
  n_cores = 1,
  keep_unmodified = FALSE
)
}
\arguments{
\item{quantification_data}{MS data, preferably in `MSstats` or `MSstatsTMT` format.}

\item{fasta_db_path}{path to a fasta file that store protein sequences}

\item{protein_column}{name of a column with protein names.}

\item{peptide_column}{name of a column with peptide sequences.}

\item{n_cores}{number of cores that will be used while searching the database.}

\item{keep_unmodified}{if TRUE, a column that stores sequences of unmodified
peptides will be added to output data}
}
\description{
Add missing proteins to PSM-level data based on a fasta database
}
