% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_checks.R
\name{checkDataCorrectness}
\alias{checkDataCorrectness}
\title{Check if data is in MSstatsTMT format}
\usage{
checkDataCorrectness(feature_data, experiment_type)
}
\description{
Check if data is in MSstatsTMT format
}
\keyword{internal}
