% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_checks.R
\name{checkExperimentType}
\alias{checkExperimentType}
\title{Check if data is label-free or TMT}
\usage{
checkExperimentType(feature_data)
}
\description{
Check if data is label-free or TMT
}
\keyword{internal}
