% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_output.r
\name{convergenceHistory,MSstatsWeightedSummary-method}
\alias{convergenceHistory,MSstatsWeightedSummary-method}
\title{Extract convergence history from MSstatsWeightedSummary object}
\usage{
\S4method{convergenceHistory}{MSstatsWeightedSummary}(weighted_summary)
}
\arguments{
\item{weighted_summary}{output of the getWeightedProteinSummary function}
}
\value{
data.table
}
\description{
Extract convergence history from MSstatsWeightedSummary object
}
