% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_output.r
\name{convergenceSummary,MSstatsWeightedSummary-method}
\alias{convergenceSummary,MSstatsWeightedSummary-method}
\title{Extract convergence information from MSstatsWeightedSummary object}
\usage{
\S4method{convergenceSummary}{MSstatsWeightedSummary}(weighted_summary)
}
\arguments{
\item{weighted_summary}{output of the getWeightedProteinSummary function}
}
\value{
data.table
}
\description{
Extract convergence information from MSstatsWeightedSummary object
}
