% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_output.r
\name{criterionValues,MSstatsWeightedSummary-method}
\alias{criterionValues,MSstatsWeightedSummary-method}
\title{Extract values of the model-fitting criterion from MSstatsWeightedSummary object}
\usage{
\S4method{criterionValues}{MSstatsWeightedSummary}(weighted_summary)
}
\arguments{
\item{weighted_summary}{output of the getWeightedProteinSummary function}
}
\value{
data.table
}
\description{
Extract values of the model-fitting criterion from MSstatsWeightedSummary object
}
