% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_output.r
\name{criterionValues}
\alias{criterionValues}
\title{Extract values of the model-fitting criterion from MSstatsWeightedSummary object}
\usage{
criterionValues(weighted_summary)
}
\arguments{
\item{weighted_summary}{output of the getWeightedProteinSummary function}
}
\description{
Extract values of the model-fitting criterion from MSstatsWeightedSummary object
}
