% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_summary.R
\name{.getDesign}
\alias{.getDesign}
\title{Get design matrix for protein-level summarization}
\usage{
.getDesign(input, design_matrix)
}
\description{
Get design matrix for protein-level summarization
}
\keyword{internal}
