% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_weights.R
\name{.getWeightsOptimProblem}
\alias{.getWeightsOptimProblem}
\title{Get optimization problem for weights estimation}
\usage{
.getWeightsOptimProblem(
  wide,
  intercept = FALSE,
  norm = "p_norm",
  norm_parameter = 1,
  weights_mode = "contributions"
)
}
\description{
Get optimization problem for weights estimation
}
\keyword{internal}
