% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_weights.R
\name{.processOptimSolution}
\alias{.processOptimSolution}
\title{Process output of optimization procedure}
\usage{
.processOptimSolution(solution, problem, num_proteins)
}
\description{
Process output of optimization procedure
}
\keyword{internal}
