% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_summary.R
\name{.processProteinSolution}
\alias{.processProteinSolution}
\title{Get protein abundances from optimization result}
\usage{
.processProteinSolution(solution, optimization_problem, input)
}
\description{
Get protein abundances from optimization result
}
\keyword{internal}
