% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_weights.R
\name{getAllWeights}
\alias{getAllWeights}
\title{Get PSM-protein weights for summarization with shared peptides}
\usage{
getAllWeights(
  input,
  norm,
  norm_parameter,
  weights_mode,
  use_control,
  control_pattern
)
}
\arguments{
\item{input}{data.table}

\item{norm}{"p_norm" or "huber"}

\item{norm_parameter}{p for norm=="p_norm", M for norm=="Huber"
that belong to a given protein will share weight parameter}

\item{weights_mode}{"contributions" for "sum to one" condition,
"probabilities" for only "non-negative" condition.}
}
\value{
data.table
}
\description{
Get PSM-protein weights for summarization with shared peptides
}
