% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_statistics.R
\name{getClusterStatistics}
\alias{getClusterStatistics}
\title{Calculate statistics that describe clusters of proteins and peptides.}
\usage{
getClusterStatistics(quantification_data, merge = FALSE)
}
\arguments{
\item{quantification_data}{MS data, preferably in `MSstats` or `MSstatsTMT` format.}

\item{merge}{if `TRUE`, calculated statistics will be merged into original data.}
}
\description{
Calculate statistics that describe clusters of proteins and peptides.
}
