% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_summary.R
\name{getClusterSummaries}
\alias{getClusterSummaries}
\title{Calculate weighted summaries for a list of clusters}
\usage{
getClusterSummaries(
  cluster_input,
  norm,
  norm_parameter,
  tolerance,
  max_iter,
  initial_summary
)
}
\arguments{
\item{cluster_input}{list of protein clusters}

\item{norm}{"p_norm" or "Huber"}

\item{norm_parameter}{p for norm=="p_norm", M for norm=="Huber"}

\item{tolerance}{tolerance to indicate weights convergence}

\item{max_iter}{maximum number of iteration of the procedure}

\item{initial_summary}{"unique", "flat" or "flat unique"}
}
\description{
Calculate weighted summaries for a list of clusters
}
\keyword{internal}
