% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_output.R
\name{getConvergenceHistory}
\alias{getConvergenceHistory}
\title{Get details of convergence}
\usage{
getConvergenceHistory(
  summary_per_cluster,
  tolerance,
  save_convergence_history,
  experiment_type
)
}
\arguments{
\item{summary_per_cluster}{output of getClusterSummaries}

\item{tolerance}{tolerance to indicate weights convergence}

\item{save_convergence_history}{logical, if TRUE, all differences between consecutive
weights estimator from all iterations will be returned}
}
\description{
Get details of convergence
}
\keyword{internal}
