% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_output.R
\name{getConvergenceSummary}
\alias{getConvergenceSummary}
\title{Get convergence summary}
\usage{
getConvergenceSummary(summary_per_cluster, tolerance, experiment_type)
}
\arguments{
\item{summary_per_cluster}{output of getClusterSummaries}

\item{tolerance}{tolerance to indicate weights convergence}
}
\description{
Get convergence summary
}
\keyword{internal}
