% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discordant_info.R
\name{getDiscordantInfo}
\alias{getDiscordantInfo}
\title{Cluster peptide profiles}
\usage{
getDiscordantInfo(input, use_control = TRUE, control_pattern = "DMSO")
}
\arguments{
\item{input}{data.table in MSstatsTMT format}

\item{use_control}{if TRUE, control group channels will be used in clustering}

\item{control_pattern}{regular expression that identifies control conditions}
}
\value{
data.table
}
\description{
Cluster peptide profiles
}
