% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_groupComparisons.R
\name{getGroupComparisons}
\alias{getGroupComparisons}
\title{Get groupComparison for different summarizations}
\usage{
getGroupComparisons(contrast_matrix, ...)
}
\arguments{
\item{contrast_matrix}{contrast matrix}

\item{...}{data.table object with summarization in MSstatsTMT format}
}
\value{
data.table
}
\description{
Get groupComparison for different summarizations
}
