% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_summary.R
\name{getInitialSummary}
\alias{getInitialSummary}
\title{Get a summary to start iterative optimization}
\usage{
getInitialSummary(
  input,
  method,
  use_discordant,
  use_shared,
  norm,
  norm_parameter
)
}
\arguments{
\item{input}{data.table}

\item{method}{if "MSstats", MSstatsTMT summarization will be used. Otherwise,
summary will be based on "norm" and "norm_parameter" parameters.}

\item{use_discordant}{if TRUE, discordant peptides will be used}

\item{use_shared}{if TRUE, shared peptides will be used}

\item{norm}{"p_norm" or "huber"}

\item{norm_parameter}{p for norm=="p_norm", M for norm=="Huber"}
}
\value{
data.table
}
\description{
Get a summary to start iterative optimization
}
