% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_summary.R
\name{getMSstatsSummary}
\alias{getMSstatsSummary}
\title{Get MSstatsTMT-based protein-level summary}
\usage{
getMSstatsSummary(input, use_shared, use_discordant)
}
\description{
Get MSstatsTMT-based protein-level summary
}
\keyword{internal}
