% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_summary.R
\name{getOptimSummary}
\alias{getOptimSummary}
\title{Optimize a give criterion to estimate protein abundances}
\usage{
getOptimSummary(
  input,
  design_type,
  use_discordant,
  norm,
  norm_parameter,
  use_shared,
  adaptive_huber
)
}
\description{
Optimize a give criterion to estimate protein abundances
}
\keyword{internal}
