% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_weights.R
\name{getPPWeights}
\alias{getPPWeights}
\title{Get weights for a single PSM}
\usage{
getPPWeights(psm_data, intercept = TRUE, norm = "p_norm", norm_parameter = 1)
}
\description{
Get weights for a single PSM
}
\keyword{internal}
