% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_checks.R
\name{getProteinsClusters}
\alias{getProteinsClusters}
\title{Split data into a list of clusters}
\usage{
getProteinsClusters(feature_data)
}
\arguments{
\item{feature_data}{data.table in MSstatsTMT format. See also the Details section}
}
\description{
Split data into a list of clusters
}
\keyword{internal}
