% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_summary.R
\name{getProteinsOptimProblem}
\alias{getProteinsOptimProblem}
\title{Get optimization problem for protein-level summaries}
\usage{
getProteinsOptimProblem(design_matrix, y, norm, norm_parameter, proteins, psms)
}
\arguments{
\item{design_matrix}{design matrix for protein model optimization}

\item{y}{observed feature intensities}

\item{norm}{"p_norm" or "Huber"}

\item{norm_parameter}{p for norm=="p_norm", M for norm=="Huber"}

\item{proteins}{vector of unique proteins}

\item{psms}{vector of unique psms}
}
\description{
Get optimization problem for protein-level summaries
}
\keyword{internal}
