% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_summary.R
\name{getProteinSummaryDesign}
\alias{getProteinSummaryDesign}
\title{Get design matrix for protein-level summarization}
\usage{
getProteinSummaryDesign(feature_data)
}
\arguments{
\item{feature_data}{data.table in MSstatsTMT format. See also the Details section}
}
\description{
Get design matrix for protein-level summarization
}
\keyword{internal}
