% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_summary.R
\name{getRobustSummary}
\alias{getRobustSummary}
\title{Get robust summary with shared peptides}
\usage{
getRobustSummary(
  input,
  method_label = "",
  norm = "p_norm",
  norm_parameter = 1,
  weights_mode = "contributions",
  adaptive_huber = FALSE,
  use_control = TRUE,
  control_pattern = NULL,
  tolerance = 0.001,
  max_iter = 10
)
}
\arguments{
\item{input}{data.table in MSstatsTMT format}

\item{method_label}{name for a summarization method}

\item{norm}{"p_norm" or "Huber"}

\item{norm_parameter}{p for norm=="p_norm", M for norm=="Huber"}

\item{weights_mode}{"contributions" for "sum to one" condition,
"probabilities" for only "non-negative" condition.}

\item{adaptive_huber}{if TRUE, M parameter for Huber norm will be updated
in each iteration of the procedure}

\item{tolerance}{tolerance to indicate weights convergence}

\item{max_iter}{maximum number of iteration of the procedure}
}
\value{
list of final summary, history of weights and difference between weights
from consecutive iterations
}
\description{
Get robust summary with shared peptides
}
