% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_summary.R
\name{getWeightedProteinSummary}
\alias{getWeightedProteinSummary}
\title{Get robust protein-level summary based on unique and shared peptides}
\usage{
getWeightedProteinSummary(
  feature_data,
  norm = "p_norm",
  norm_parameter = 1,
  weights_mode = "contributions",
  tolerance = 0.1,
  max_iter = 10,
  initial_summary = "unique",
  weights_penalty = FALSE,
  weights_penalty_param = 0.1,
  save_weights_history = FALSE,
  save_convergence_history = FALSE
)
}
\arguments{
\item{feature_data}{data.table in MSstatsTMT format. See also the Details section}

\item{norm}{"p_norm" or "Huber"}

\item{norm_parameter}{p for norm=="p_norm", M for norm=="Huber"}

\item{weights_mode}{"contributions" for "sum to one" and "non-negative" conditions,
"probabilities" for only "non-negative" condition.}

\item{tolerance}{tolerance to indicate weights convergence}

\item{max_iter}{maximum number of iteration of the procedure}

\item{initial_summary}{"unique", "flat" or "flat unique"}

\item{weights_penalty}{if TRUE, weights will be penalized for deviations from equal value
for all proteins matching to a given PSM}

\item{weights_penalty_param}{penalty parameter}

\item{save_weights_history}{logical, if TRUE, weights from all iterations will
be returned}

\item{save_convergence_history}{logical, if TRUE, all differences between consecutive
weights estimator from all iterations will be returned}
}
\value{
list of data frames with summary and other information. See the Details section for more information
}
\description{
Get robust protein-level summary based on unique and shared peptides
}
\details{
1. Input format: this function takes as input data in MSstatsTMT format,
which is a data frame with columns ProteinName, PeptideSequence, Charge,
PSM (equal to PeptideSequence and Charge separated by an underscore),
Channel, Intensity, Run and annotation columns: BioReplicate, Condition,
Mixture, and TechRepMixture. Additionally, we use two columns:
log2IntensityNormalized and Cluster. The first column stores log-transformed
normalized intensities (which can be obtained with \code{\link{normalizeSharedPeptides}} function).
If this column is not provided, data will be normalized before summarization.
The second column stores information about connected sub-graphs of the
peptide-protein graph. This column can be added with \code{\link{addClusterMembership}} function or
omitted. In the second case, this information will be added before summarization.

2. Output format: an S4 object of class "MSstatsWeightedSummary" which consists of the following items:
\itemize{
\item{FeatureLevelData:}{feature-level (input) data}
\item{ProteinLevelData:}{protein-level (summarized) output data}
\item{Weights:}{a table of final peptide-protein Weights}
\item{ConvergenceSummary:}{table with information about convergence for each Cluster and Run}
\item{WeightsHistory:}{optional data.table of Weights from all iterations of fitting algorithm}
\item{ConvergenceHistory:}{optional data.table with sums of absolute values of differences between Weights from consecutive iteration}
}
Elements of this object can be accessed with functions
\code{\link{featureData}}, \code{\link{proteinData}},
\code{\link{featureWeights}}, \code{\link{convergenceSummary}},
\code{\link{weightsHistory}}, \code{\link{convergenceHistory}}

For statistical details about the method, please consult the vignette.
}
