% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_summary.R
\name{getWeightedSummary}
\alias{getWeightedSummary}
\title{Optimize a given criterion to estimate protein abundances}
\usage{
getWeightedSummary(
  input,
  weights_df,
  norm,
  norm_parameter,
  use_shared,
  adaptive_huber
)
}
\arguments{
\item{input}{data.table in MSstatsTMT format}

\item{weights_df}{data.table with weights.}

\item{norm}{"p_norm" or "Huber"}

\item{norm_parameter}{p for norm=="p_norm", M for norm=="Huber"}

\item{use_shared}{if TRUE, shared peptides will be used.}

\item{adaptive_huber}{if TRUE, M parameter for Huber norm will be updated
in each iteration of the procedure}
}
\description{
Optimize a given criterion to estimate protein abundances
}
