% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_summary.R
\name{getWeightedSummarySingleRun}
\alias{getWeightedSummarySingleRun}
\title{Robust summary for a single run}
\usage{
getWeightedSummarySingleRun(
  feature_data,
  peptide_protein_dt,
  norm,
  norm_parameter,
  weights_mode,
  tolerance,
  max_iter,
  initial_summary,
  weights_penalty,
  weights_penalty_param
)
}
\arguments{
\item{feature_data}{data.table in MSstatsTMT format. See also the Details section}

\item{peptide_protein_dt}{table of peptide-protein matches}

\item{norm}{"p_norm" or "Huber"}

\item{norm_parameter}{p for norm=="p_norm", M for norm=="Huber"}

\item{tolerance}{tolerance to indicate weights convergence}

\item{max_iter}{maximum number of iteration of the procedure}

\item{initial_summary}{"unique", "flat" or "flat unique"}
}
\description{
Robust summary for a single run
}
\keyword{internal}
