% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_weights.R
\name{getWeightsDesign}
\alias{getWeightsDesign}
\title{Get design matrix and response for weights calculation}
\usage{
getWeightsDesign(feature_data)
}
\arguments{
\item{feature_data}{data.table in MSstatsTMT format. See also the Details section}
}
\description{
Get design matrix and response for weights calculation
}
\keyword{internal}
