% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_output.R
\name{getWeightsSummary}
\alias{getWeightsSummary}
\title{Get summary of final weights}
\usage{
getWeightsSummary(summary_per_cluster, experiment_type)
}
\arguments{
\item{summary_per_cluster}{output of getClusterSummaries}
}
\description{
Get summary of final weights
}
\keyword{internal}
