% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_output.r
\name{makeMSstatsInput,MSstatsWeightedSummary-method}
\alias{makeMSstatsInput,MSstatsWeightedSummary-method}
\title{Create input for MSstats::groupComparison function}
\usage{
\S4method{makeMSstatsInput}{MSstatsWeightedSummary}(weighted_summary, msstats_output = NULL)
}
\arguments{
\item{weighted_summary}{output of the getWeightedProteinSummary function}

\item{msstatstmt_output}{optional output of MSstats::dataProcess function}
}
\value{
list
}
\description{
Create input for MSstats::groupComparison function
}
