% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_output.r
\name{makeMSstatsInput}
\alias{makeMSstatsInput}
\title{Create input for MSstats::groupComparison function}
\usage{
makeMSstatsInput(weighted_summary, msstats_output = NULL)
}
\arguments{
\item{weighted_summary}{output of the getWeightedProteinSummary function}

\item{msstats_output}{optional output of MSstats::dataProcess function}
}
\description{
Create input for MSstats::groupComparison function
}
