% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_output.r
\name{makeMSstatsTMTInput}
\alias{makeMSstatsTMTInput}
\title{Create input for MSstatsTMT::groupComparisonTMT function}
\usage{
makeMSstatsTMTInput(weighted_summary, msstatstmt_output = NULL)
}
\arguments{
\item{weighted_summary}{output of the getWeightedProteinSummary function}

\item{msstatstmt_output}{optional output of MSstatsTMT::proteinSummarization function}
}
\description{
Create input for MSstatsTMT::groupComparisonTMT function
}
