% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_normalize.R
\name{normalizePeptides}
\alias{normalizePeptides}
\title{Normalization between channels (before summarization)}
\usage{
normalizePeptides(feature_data)
}
\arguments{
\item{feature_data}{data.table in MSstatsTMT format. See also the Details section}
}
\value{
data.table
}
\description{
Normalization between channels (before summarization)
}
\author{
based on MSstatsTMT code by Ting Huang
}
\keyword{internal}
