% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_normalize.R
\name{normalizeSharedPeptides}
\alias{normalizeSharedPeptides}
\title{Normalization for TMT data}
\usage{
normalizeSharedPeptides(input)
}
\arguments{
\item{input}{data.table}
}
\value{
data.table
}
\description{
Normalization for TMT data
}
