% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_statistics.R
\name{plotClusterStats}
\alias{plotClusterStats}
\title{Plot statistics about clusters of proteins}
\usage{
plotClusterStats(
  cluster_statistics,
  statistic,
  only_clusters_with_shared = FALSE
)
}
\arguments{
\item{cluster_statistics}{output of the `getClusterStatistics` function.}

\item{statistic}{one of the column names of the `cluster_statistics` parameter.}

\item{only_clusters_with_shared}{if `TRUE`, statistics will be presented only
for cluster with more than one protein.}
}
\description{
Plot statistics about clusters of proteins
}
