% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plots.R
\name{plotDistributionComparison}
\alias{plotDistributionComparison}
\title{Compare distribution of log2-intensities per channel with and without shared peptides}
\usage{
plotDistributionComparison(input, plot_profiles = FALSE)
}
\arguments{
\item{input}{data.table}

\item{plot_profiles}{if TRUE, feature profiles will be plotted}
}
\value{
ggplot
}
\description{
Compare distribution of log2-intensities per channel with and without shared peptides
}
