% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plots.R
\name{plotSummarizedProteins}
\alias{plotSummarizedProteins}
\title{Plot protein-level summaries for a selected cluster of proteins}
\usage{
plotSummarizedProteins(weighted_summary, cluster, channel_order = NULL)
}
\arguments{
\item{weighted_summary}{Output of the getWeightedProteinSummary function}

\item{cluster}{ID of a cluster of proteins summarized in the `weighted_summary` object}

\item{channel_order}{optional character vector of labels of Runs or Channels which
will be used to sort the x-axis}
}
\value{
ggplot2 object
}
\description{
Plot protein-level summaries for a selected cluster of proteins
}
