% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_output.R
\name{plotSummary,MSstatsWeightedSummary-method}
\alias{plotSummary,MSstatsWeightedSummary-method}
\title{Plot summary for a given cluster or proteins set}
\usage{
\S4method{plotSummary}{MSstatsWeightedSummary}(
  weighted_summary,
  cluster = NULL,
  proteins = NULL,
  channel_order = NULL
)
}
\arguments{
\item{weighted_summary}{output of the getWeightedProteinSummary function}

\item{cluster}{optional ID of a cluster to plot. Either `cluster` or `proteins` must be provided}

\item{proteins}{optional vector of proteins to plot. Either `cluster` or `proteins` must be provided}

\item{channel_order}{optional vector of ordered channel IDs. If provided,
x-axis of the plot will follow this order}
}
\value{
ggplot2 object
}
\description{
Plot summary for a given cluster or proteins set
}
