% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_output.R
\name{plotSummary}
\alias{plotSummary}
\title{Plot summary for a given cluster or proteins set}
\usage{
plotSummary(
  weighted_summary,
  cluster = NULL,
  proteins = NULL,
  channel_order = NULL
)
}
\value{
ggplot2 object
}
\description{
Plot summary for a given cluster or proteins set
}
