% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_isoforms.R
\name{processIsoforms}
\alias{processIsoforms}
\title{Process protein isoforms to prepare data for MSstats analysis}
\usage{
processIsoforms(
  quantification_data,
  remove_single_shared = TRUE,
  merge_identical = TRUE,
  remove_subsets = FALSE,
  subset_treatment = "remove_all"
)
}
\arguments{
\item{quantification_data}{data in MSstatsTMT format}

\item{remove_single_shared}{if TRUE (default), proteins identified by a single shared peptide will be removed}

\item{merge_identical}{if TRUE (default), proteins identified by the same set of peptides in a run will be merged
into a single protein label (concatenation of their IDs)}

\item{remove_subsets}{if TRUE, proteins with no unique peptides in a run will be removed from the analysis (defaults to FALSE)}

\item{subset_treatment}{character describing the treatment of sets of proteins that do not have unique peptides.
Currently, only removing or keeping all subset proteins is possible.}
}
\value{
data.table in a format compatible with MSstatsWeightedSummary workflow
}
\description{
Process protein isoforms to prepare data for MSstats analysis
}
