% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_summary.R
\name{processProteinOptimSolution}
\alias{processProteinOptimSolution}
\title{Get protein abundances from optimization result}
\usage{
processProteinOptimSolution(
  solution,
  optimization_problem,
  feature_data,
  design_matrix
)
}
\arguments{
\item{solution}{solution generated by CVXR::solve}

\item{optimization_problem}{object returned by CVXR::Problem}

\item{feature_data}{data.table in MSstatsTMT format. See also the Details section}

\item{design_matrix}{design matrix for protein model optimization}
}
\description{
Get protein abundances from optimization result
}
\keyword{internal}
