% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_output.r
\name{proteinClusters,MSstatsWeightedSummary-method}
\alias{proteinClusters,MSstatsWeightedSummary-method}
\title{Extract cluster information from MSstatsWeightedSummary object}
\usage{
\S4method{proteinClusters}{MSstatsWeightedSummary}(weighted_summary)
}
\arguments{
\item{weighted_summary}{output of the getWeightedProteinSummary function}
}
\value{
data.table
}
\description{
Extract cluster information from MSstatsWeightedSummary object
}
