% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_output.r
\name{proteinData}
\alias{proteinData}
\title{Extract protein-level data from MSstatsWeightedSummary object}
\usage{
proteinData(weighted_summary, proteins = NULL)
}
\arguments{
\item{weighted_summary}{output of the getWeightedProteinSummary function}

\item{proteins}{optional character vector of proteins to extract. If NULL, all proteins
will be returned}
}
\description{
Extract protein-level data from MSstatsWeightedSummary object
}
