% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSstatsWeightedSummary-package.R
\docType{data}
\name{simulated_dataset}
\alias{simulated_dataset}
\title{Simulated data set illustrating quantification with shared peptides}
\format{
A data frame (data.table) with 168 rows and 14 variables:
\describe{
  \item{ProteinName}{label of protein matching feature's AA sequence}
  \item{PeptideSequence}{AA sequence of the feature}
  \item{Charge}{charge of the feature)}
  \item{PSM}{PSM label - concatenation of PeptideSequence and Charge columns}
  \item{Run}{label of MS run (constant)}
  \item{Mixture}{label of a TMT mixture (constant)}
  \item{TechRepMixture}{label of technical replicate mixture (constant)}
  \item{Channel}{label of a TMT channel}
  \item{BioReplicate}{label of biological replicate (group comparison design)}
  \item{Condition}{label of a group}
  \item{Log2Intensity}{log-intensity values}
  \item{Intensity}{intensity values}
  \item{IsUnique}{logical column indicating if a given peptide is unique to its parent protein}
  \item{log2IntensityNormalized}{normalized log-intensities of features}
}
}
\usage{
simulated_dataset
}
\description{
Quantification data in the MSstatsTMT format with characteristics similar to the
BRD case study presented in the original publication.
}
\keyword{datasets}
