% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_summary.R
\name{summarizeProteinsClusterSingleRun}
\alias{summarizeProteinsClusterSingleRun}
\title{Optimize a given criterion to estimate protein abundances}
\usage{
summarizeProteinsClusterSingleRun(
  feature_data,
  weights,
  norm,
  norm_parameter,
  use_shared
)
}
\arguments{
\item{feature_data}{data.table in MSstatsTMT format. See also the Details section}

\item{weights}{data.table with weights.}

\item{norm}{"p_norm" or "Huber"}

\item{norm_parameter}{p for norm=="p_norm", M for norm=="Huber"}

\item{use_shared}{if TRUE, shared peptides will be used.}
}
\description{
Optimize a given criterion to estimate protein abundances
}
