% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_output.r
\name{weightsHistory}
\alias{weightsHistory}
\title{Extract weights history from MSstatsWeightedSummary object}
\usage{
weightsHistory(weighted_summary, shared_only = TRUE)
}
\arguments{
\item{weighted_summary}{output of the getWeightedProteinSummary function}

\item{shared_only}{logical, output data.table will only include shared peptides}
}
\description{
Extract weights history from MSstatsWeightedSummary object
}
