% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_output.r
\name{convergenceHistory}
\alias{convergenceHistory}
\title{Extract convergence history from MSstatsWeightedSummary object}
\usage{
convergenceHistory(weighted_summary)
}
\arguments{
\item{weighted_summary}{output of the getWeightedProteinSummary function}
}
\description{
Extract convergence history from MSstatsWeightedSummary object
}
