% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_output.r
\name{convergenceSummary}
\alias{convergenceSummary}
\title{Extract convergence information from MSstatsWeightedSummary object}
\usage{
convergenceSummary(weighted_summary)
}
\description{
Extract convergence information from MSstatsWeightedSummary object
}
