% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_statistics.R
\name{createPeptideProteinGraph}
\alias{createPeptideProteinGraph}
\title{Wrapper around `igraph::graph_from_data_frame`}
\usage{
createPeptideProteinGraph(
  quantification_data,
  protein_column = "ProteinName",
  peptide_column = "PeptideSequence"
)
}
\arguments{
\item{quantification_data}{MS data, preferably in `MSstats` or `MSstatsTMT` format.}

\item{protein_column}{name of a column with protein names.}

\item{peptide_column}{name of a column with peptide sequences.}
}
\description{
Wrapper around `igraph::graph_from_data_frame`
}
