% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_summary.R
\name{.getProteinsOptimProblem}
\alias{.getProteinsOptimProblem}
\title{Get optimization problem for protein-level summaries}
\usage{
.getProteinsOptimProblem(design_matrix, y, norm = "p_norm", norm_parameter = 1)
}
\description{
Get optimization problem for protein-level summaries
}
\keyword{internal}
