% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_weights.R
\name{.getWeightsCombined}
\alias{.getWeightsCombined}
\title{Get weights based on joint optimization (not per PSM)}
\usage{
.getWeightsCombined(
  input,
  norm = "p_norm",
  norm_parameter = 1,
  weights_mode = "contributions",
  use_control,
  control_pattern
)
}
\description{
Get weights based on joint optimization (not per PSM)
}
\keyword{internal}
