% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_output.R
\name{featureData,MSstatsWeightedSummary-method}
\alias{featureData,MSstatsWeightedSummary-method}
\title{Extract feature-level data from MSstatsWeightedSummary object}
\usage{
\S4method{featureData}{MSstatsWeightedSummary}(weighted_summary, proteins = NULL)
}
\arguments{
\item{weighted_summary}{output of the getWeightedProteinSummary function}

\item{proteins}{optional character vector of proteins to extract. If NULL, all proteins
will be returned}
}
\value{
data.table
}
\description{
Extract feature-level data from MSstatsWeightedSummary object
}
