% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_output.R
\name{featureData}
\alias{featureData}
\title{Extract feature-level data from MSstatsWeightedSummary object}
\usage{
featureData(weighted_summary, proteins = NULL)
}
\description{
Extract feature-level data from MSstatsWeightedSummary object
}
