% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_output.R
\name{featureWeights,MSstatsWeightedSummary-method}
\alias{featureWeights,MSstatsWeightedSummary-method}
\title{Extract weights data from MSstatsWeightedSummary object}
\usage{
\S4method{featureWeights}{MSstatsWeightedSummary}(weighted_summary, proteins = NULL, shared_only = TRUE)
}
\arguments{
\item{weighted_summary}{output of the getWeightedProteinSummary function}

\item{proteins}{optional character vector of proteins to extract. If NULL, all proteins
will be returned}

\item{shared_only}{logical, output data.table will only include shared peptides}
}
\value{
data.table
}
\description{
Extract weights data from MSstatsWeightedSummary object
}
