% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_output.R
\name{featureWeights}
\alias{featureWeights}
\title{Extract weights data from MSstatsWeightedSummary object}
\usage{
featureWeights(weighted_summary, proteins = NULL, shared_only = TRUE)
}
\description{
Extract weights data from MSstatsWeightedSummary object
}
