% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_output.R
\name{fittedProfiles,MSstatsWeightedSummary-method}
\alias{fittedProfiles,MSstatsWeightedSummary-method}
\title{Extract cluster information from MSstatsWeightedSummary object}
\usage{
\S4method{fittedProfiles}{MSstatsWeightedSummary}(weighted_summary)
}
\arguments{
\item{weighted_summary}{output of the getWeightedProteinSummary function}
}
\value{
data.table
}
\description{
Extract cluster information from MSstatsWeightedSummary object
}
