% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_output.R
\name{fittedProfiles}
\alias{fittedProfiles}
\title{Extract fitted PSM-level profiles from MSstatsWeightedSummary object}
\usage{
fittedProfiles(weighted_summary)
}
\description{
Extract fitted PSM-level profiles from MSstatsWeightedSummary object
}
