% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_summary.R
\name{getCurrentWeights}
\alias{getCurrentWeights}
\title{Utility function to get weights from current iterations including 0s (removed features)}
\usage{
getCurrentWeights(weights, peptide_protein_dt)
}
\arguments{
\item{weights}{table of weights (columns ProteinName, PSM, Weight)}

\item{peptide_protein_dt}{table of peptide-protein matches}
}
\description{
Utility function to get weights from current iterations including 0s (removed features)
}
\keyword{internal}
