% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_summary.R
\name{getRobustSummary}
\alias{getRobustSummary}
\title{Get robust summary with shared peptides}
\usage{
getRobustSummary(
  input,
  method_label,
  design_type = "short",
  norm = "p_norm",
  norm_parameter = 1,
  use_weights = TRUE,
  weights_method = "separate",
  adaptive_huber = TRUE,
  initial_summary_method = "msstats",
  use_discordant = TRUE,
  use_shared_initial = FALSE,
  tolerance = 0.001,
  max_iter = 10
)
}
\arguments{
\item{input}{data.table in MSstatsTMT format}

\item{method_label}{name for a summarization method}

\item{design_type}{"short" or "long" (not implemented yet)}

\item{norm}{"p_norm" or "Huber"}

\item{norm_parameter}{p for norm=="p_norm", M for norm=="Huber"}

\item{use_weights}{if TRUE, PSM-Protein weights will be used}

\item{weights_method}{if "separate", weights will be estimated for each PSM
separately, if "all", weights will be estimated jointly}

\item{adaptive_huber}{if TRUE, M parameter for Huber norm will be updated
in each iteration of the procedure}

\item{initial_summary_method}{summarization method to start iterative procedure
- "msstats" or "same"}

\item{use_discordant}{if TRUE, discordant profiles will be used}

\item{use_shared_initial}{if TRUE, shared peptides will be used in summary
that initializes iterative procedure}

\item{tolerance}{tolerance to indicate weights convergence}

\item{max_iter}{maximum number of iteration of the procedure}
}
\value{
list of final summary, history of weights and difference between weights
from consecutive iterations
}
\description{
Get robust summary with shared peptides
}
