% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_summary.R
\name{getRobustSummarySingleRun}
\alias{getRobustSummarySingleRun}
\title{Robust summary for a single run}
\usage{
getRobustSummarySingleRun(
  input,
  method_label,
  norm = "p_norm",
  norm_parameter = 1,
  weights_mode = "contributions",
  adaptive_huber = FALSE,
  use_control = TRUE,
  control_pattern = NULL,
  tolerance = 0.001,
  max_iter = 10
)
}
\description{
Robust summary for a single run
}
\keyword{internal}
