% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_summary.R
\name{getWeightedSummarySingleRun}
\alias{getWeightedSummarySingleRun}
\title{Robust summary for a single run}
\usage{
getWeightedSummarySingleRun(
  feature_data,
  peptide_protein_dt,
  norm,
  norm_parameter,
  weights_mode,
  tolerance,
  max_iter
)
}
\arguments{
\item{feature_data}{data.table in MSstatsTMT format. See also the Details section}

\item{peptide_protein_dt}{table of peptide-protein matches}

\item{norm}{"p_norm" or "Huber"}

\item{norm_parameter}{p for norm=="p_norm", M for norm=="Huber"}

\item{weights_mode}{"contributions" for "sum to one" and "non-negative" conditions,
"probabilities" for only "non-negative" condition.}

\item{tolerance}{tolerance to indicate weights convergence}

\item{max_iter}{maximum number of iteration of the procedure}
}
\description{
Robust summary for a single run
}
\keyword{internal}
