% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_weights.R
\name{getWeightsConstraints}
\alias{getWeightsConstraints}
\title{Get constraints for weights}
\usage{
getWeightsConstraints(
  params_full,
  design_matrix,
  weights_mode,
  cols,
  protein_cols,
  psms_cols
)
}
\arguments{
\item{params_full}{object generated by CVXR::Variables}

\item{design_matrix}{design matrix for weights optimization}

\item{weights_mode}{"contributions" for "sum to one" and "non-negative" conditions,
"probabilities" for only "non-negative" condition.}

\item{cols}{column names of the design matrix}

\item{protein_cols}{protein names extracted from column names}

\item{psms_cols}{feature names extracted from column names}
}
\description{
Get constraints for weights
}
