% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_output.R
\name{getWeightsHistory}
\alias{getWeightsHistory}
\title{Get history of weights from all iterations}
\usage{
getWeightsHistory(summary_per_cluster, save_weights_history)
}
\arguments{
\item{summary_per_cluster}{output of getClusterSummaries}

\item{save_weights_history}{logical, if TRUE, weights from all iterations will
be returned}
}
\description{
Get history of weights from all iterations
}
\keyword{internal}
