% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_output.R
\name{makeMSstatsInput}
\alias{makeMSstatsInput}
\title{Create input for MSstats::groupComparison function}
\usage{
makeMSstatsInput(weighted_summary, msstats_output = NULL)
}
\description{
Create input for MSstats::groupComparison function
}
