% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_output.R
\name{makeMSstatsTMTInput,MSstatsWeightedSummary-method}
\alias{makeMSstatsTMTInput,MSstatsWeightedSummary-method}
\title{Create input for MSstatsTMT::groupComparison function}
\usage{
\S4method{makeMSstatsTMTInput}{MSstatsWeightedSummary}(weighted_summary, msstatstmt_output = NULL)
}
\arguments{
\item{weighted_summary}{output of the getWeightedProteinSummary function}

\item{msstatstmt_output}{optional output of MSstatsTMT::proteinSummarization function}
}
\value{
list
}
\description{
Create input for MSstatsTMT::groupComparison function
}
