% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_output.r
\name{makeMSstatsTMTInput}
\alias{makeMSstatsTMTInput}
\title{Create input for MSstatsTMT::groupComparisonTMT function}
\usage{
makeMSstatsTMTInput(weighted_summary, msstatstmt_output = NULL)
}
\description{
Create input for MSstatsTMT::groupComparisonTMT function
}
