% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_output.R
\name{makeMSstatsTMTInput}
\alias{makeMSstatsTMTInput}
\title{Create input for MSstatsTMT::groupComparison function}
\usage{
makeMSstatsTMTInput(weighted_summary, msstatstmt_output = NULL)
}
\description{
Create input for MSstatsTMT::groupComparison function
}
