% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_output.R
\docType{class}
\name{MSstatsWeightedSummary-class}
\alias{MSstatsWeightedSummary-class}
\title{Output of weighted summarization}
\description{
Output of weighted summarization
}
\section{Slots}{

\describe{
\item{\code{FeatureLevelData}}{feature-level (input) data}

\item{\code{ProteinLevelData}}{protein-level (summarized) output data}

\item{\code{Weights}}{a table of final peptide-protein Weights}

\item{\code{ConvergenceSummary}}{table with information about convergence for each Cluster and Run}

\item{\code{WeightsHistory}}{optional data.table of Weights from all iterations of fitting algorithm}

\item{\code{ConvergenceHistory}}{optional data.table with sums of absolute values of differences between Weights from consecutive iteration}
}}

