% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSstatsWeightedSummary-package.R
\docType{package}
\name{MSstatsWeightedSummary-package}
\alias{MSstatsWeightedSummary}
\alias{MSstatsWeightedSummary-package}
\title{MSstatsWeightedSummary: Weighted Protein-Level Summarization for Protein Clusters}
\description{
MSstatsWeightedSummary implements a statistical model for joint analysis of proteins that share peptides based on mass spectrometry data.
The package extends existing MSstats and MSstatsTMT workflows to include shared peptides in input data for protein summarization.
The summarization workflow is implemented in the \code{\link{getWeightedProteinSummary}} function. For example, check the package vignette
`browseVignettes(package = "MSstatsWeightedSummary.")`.
For the statistical details and background, please refer to the publication \link{https://doi.org/10.1093/bioinformatics/btaf046}.
}
\author{
\strong{Maintainer}: Mateusz Staniak \email{mtst@mstaniak.pl}

Authors:
\itemize{
  \item Małgorzata Bogdan \email{malgorzata.bogdan@uwr.edu.pl}
  \item Olga Vitek \email{o.vitek@northeastern.edu}
}

Other contributors:
\itemize{
  \item Ting Huang \email{thuang0703@gmail.com} [contributor]
}

}
\keyword{internal}
