% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_normalize.R
\name{normalizeSharedPeptides}
\alias{normalizeSharedPeptides}
\title{Normalization for TMT data with shared peptides}
\usage{
normalizeSharedPeptides(feature_data)
}
\arguments{
\item{feature_data}{data.table in MSstatsTMT format. See also the Details section}
}
\value{
data.table
}
\description{
Normalization for TMT data with shared peptides
}
