% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plots.R
\name{plotSummaryComparison}
\alias{plotSummaryComparison}
\title{Plot multiple protein-level summaries}
\usage{
plotSummaryComparison(..., channel_order = NULL, feature_data = NULL)
}
\arguments{
\item{...}{data.tables with summaries}

\item{channel_order}{optional order for x-axis (Channel column)}

\item{feature_data}{optional data.table for plotting feature-level distribution
in each channel}
}
\value{
ggplot
}
\description{
Plot multiple protein-level summaries
}
