% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_output.R
\name{processSummarizationOutput}
\alias{processSummarizationOutput}
\title{Prepare summarizaton output}
\usage{
processSummarizationOutput(
  summary_per_cluster,
  feature_data,
  annotation,
  save_weights_history,
  save_convergence_history,
  tolerance
)
}
\arguments{
\item{summary_per_cluster}{output of getClusterSummaries}

\item{feature_data}{data.table in MSstatsTMT format. See also the Details section}

\item{annotation}{output of getAnnotation}

\item{save_weights_history}{logical, if TRUE, weights from all iterations will
be returned}

\item{save_convergence_history}{logical, if TRUE, all differences between consecutive
weights estimator from all iterations will be returned}

\item{tolerance}{tolerance to indicate weights convergence}
}
\description{
Prepare summarizaton output
}
\keyword{internal}
