% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_output.r
\name{proteinClusters}
\alias{proteinClusters}
\title{Extract cluster information from MSstatsWeightedSummary object}
\usage{
proteinClusters(weighted_summary)
}
\arguments{
\item{weighted_summary}{output of the getWeightedProteinSummary function}
}
\description{
Extract cluster information from MSstatsWeightedSummary object
}
