% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_output.R
\name{proteinClusters}
\alias{proteinClusters}
\title{Extract cluster information from MSstatsWeightedSummary object}
\usage{
proteinClusters(weighted_summary)
}
\description{
Extract cluster information from MSstatsWeightedSummary object
}
