% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_output.R
\name{proteinData}
\alias{proteinData}
\title{Extract protein-level data from MSstatsWeightedSummary object}
\usage{
proteinData(weighted_summary, proteins = NULL)
}
\description{
Extract protein-level data from MSstatsWeightedSummary object
}
