% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_output.R
\name{weightsHistory,MSstatsWeightedSummary-method}
\alias{weightsHistory,MSstatsWeightedSummary-method}
\title{Extract weights history from MSstatsWeightedSummary object}
\usage{
\S4method{weightsHistory}{MSstatsWeightedSummary}(weighted_summary, shared_only = TRUE)
}
\arguments{
\item{weighted_summary}{output of the getWeightedProteinSummary function}

\item{shared_only}{logical, output data.table will only include shared peptides}
}
\value{
data.table
}
\description{
Extract weights history from MSstatsWeightedSummary object
}
