% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_output.r
\name{weightsHistory}
\alias{weightsHistory}
\title{Extract weights history from MSstatsWeightedSummary object}
\usage{
weightsHistory(weighted_summary, shared_only = TRUE)
}
\description{
Extract weights history from MSstatsWeightedSummary object
}
